feat: add button to clear filters when applied filters returns no results
