+++ src/components/ui/MovieGrid.tsx
@@ -3,7 +3,7 @@ import "@styles/MovieCard.css"
 import MovieCard from "@components/ui/MovieCard";
 import { MovieList, MovieInfo, MovieFilters, MovieSortMethod } from "types";
 import { SetStateAction, useEffect } from "react";
-import { checkIfFiltered, clearAllFilters } from "@components/utils/MovieFilterUtils";
+import { checkIfFiltered, resetAllFilters } from "@components/utils/MovieFilterUtils";
 import { useMovieSearchRef } from "@components/contexts/MovieSearchContext";
 import { Search } from "lucide-react";
 
@@ -107,7 +107,7 @@ function MovieGrid({
             </h3>
             <button 
             className="no-movies-search-button" 
-            onClick={() => {setMovieFilters(clearAllFilters(currentMovieFilters))}}>Clear Filters</button>
+            onClick={() => {setMovieFilters(resetAllFilters(currentMovieFilters)); console.log(currentMovieFilters); }}>Clear Filters</button>
           </div>
         </div>
         )
